filetr  market Intl

+++ src/main/java/repository/mapper/LeadRowMapper.java
@@ -22,7 +22,7 @@ public class LeadRowMapper implements RowMapper<LeadEntity> {
 		leadEntity.setUpdateDate(rs.getDate("UPDATE_DATE"));
 		leadEntity.setCreatorId(rs.getLong("CREATOR_ID"));
 		leadEntity.setUpdatorId(rs.getLong("UPDATOR_ID"));
-		leadEntity.setBudget(rs.getLong("BUDGET"));
+		leadEntity.setBudget(rs.getDouble("BUDGET"));
 		leadEntity.setCurrency(rs.getString("CURRENCY"));
 		leadEntity.setMessage(rs.getString("MESSAGE"));
 		return leadEntity;
